chore: split column and row pinning
