+++ packages/table-core/src/features/ColumnPinning.ts
@@ -11,24 +11,14 @@ import {
 import { getMemoOptions, makeStateUpdater, memo } from '../utils'
 
 export type ColumnPinningPosition = false | 'left' | 'right'
-export type RowPinningPosition = false | 'top' | 'bottom'
 
 export interface ColumnPinningState {
   left?: string[]
   right?: string[]
 }
 
-export interface RowPinningState {
-  bottom?: string[]
-  top?: string[]
-}
-
 export interface ColumnPinningTableState {
   columnPinning: ColumnPinningState
-}
-
-export interface RowPinningTableState {
-  rowPinning: RowPinningState
 }
 
 export interface ColumnPinningOptions {
@@ -39,6 +29,7 @@ export interface ColumnPinningOptions {
    */
   enableColumnPinning?: boolean
   /**
+   * @deprecated Use `enableColumnPinning` or `enableRowPinning` instead.
    * Enables/disables all pinning for the table. Defaults to `true`.
    * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#enablepinning)
    * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
@@ -52,33 +43,8 @@ export interface ColumnPinningOptions {
   onColumnPinningChange?: OnChangeFn<ColumnPinningState>
 }
 
-export interface RowPinningOptions<TData extends RowData> {
-  /**
-   * Enables/disables row pinning for the table. Defaults to `true`.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#enablerowpinning)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  enableRowPinning?: boolean | ((row: Row<TData>) => boolean)
-  /**
-   * When `false`, pinned rows will not be visible if they are filtered or paginated out of the table. When `true`, pinned rows will always be visible regardless of filtering or pagination. Defaults to `true`.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#keeppinnedrows)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  keepPinnedRows?: boolean
-  /**
-   * If provided, this function will be called with an `updaterFn` when `state.rowPinning` changes. This overrides the default internal state management, so you will also need to supply `state.rowPinning` from your own managed state.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#onrowpinningchange)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/onrowpinningchange)
-   */
-  onRowPinningChange?: OnChangeFn<RowPinningState>
-}
-
 export interface ColumnPinningDefaultOptions {
   onColumnPinningChange: OnChangeFn<ColumnPinningState>
-}
-
-export interface RowPinningDefaultOptions {
-  onRowPinningChange: OnChangeFn<RowPinningState>
 }
 
 export interface ColumnPinningColumnDef {
@@ -138,37 +104,6 @@ export interface ColumnPinningRow<TData extends RowData> {
   getRightVisibleCells: () => Cell<TData, unknown>[]
 }
 
-export interface RowPinningRow {
-  /**
-   * Returns whether or not the row can be pinned.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#getcanpin-1)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  getCanPin: () => boolean
-  /**
-   * Returns the pinned position of the row. (`'top'`, `'bottom'` or `false`)
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#getispinned-1)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  getIsPinned: () => RowPinningPosition
-  /**
-   * Returns the numeric pinned index of the row within a pinned row group.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#getpinnedindex-1)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  getPinnedIndex: () => number
-  /**
-   * Pins a row to the `'top'` or `'bottom'`, or unpins the row to the center if `false` is passed.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#pin-1)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  pin: (
-    position: RowPinningPosition,
-    includeLeafRows?: boolean,
-    includeParentRows?: boolean
-  ) => void
-}
-
 export interface ColumnPinningInstance<TData extends RowData> {
   /**
    * Returns all center pinned (unpinned) leaf columns.
@@ -208,46 +143,6 @@ export interface ColumnPinningInstance<TData extends RowData> {
   setColumnPinning: (updater: Updater<ColumnPinningState>) => void
 }
 
-export interface RowPinningInstance<TData extends RowData> {
-  _getPinnedRows: (position: 'top' | 'bottom') => Row<TData>[]
-  /**
-   * Returns all bottom pinned rows.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#getbottomrows)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  getBottomRows: () => Row<TData>[]
-  /**
-   * Returns all rows that are not pinned to the top or bottom.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#getcenterrows)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  getCenterRows: () => Row<TData>[]
-  /**
-   * Returns whether or not any rows are pinned. Optionally specify to only check for pinned rows in either the `top` or `bottom` position.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#getissomerowspinned)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  getIsSomeRowsPinned: (position?: RowPinningPosition) => boolean
-  /**
-   * Returns all top pinned rows.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#gettoprows)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  getTopRows: () => Row<TData>[]
-  /**
-   * Resets the **rowPinning** state to `initialState.rowPinning`, or `true` can be passed to force a default blank state reset to `{ top: [], bottom: [], }`.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#resetrowpinning)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  resetRowPinning: (defaultState?: boolean) => void
-  /**
-   * Sets or updates the `state.rowPinning` state.
-   * @link [API Docs](https://tanstack.com/table/v8/docs/api/features/pinning#setrowpinning)
-   * @link [Guide](https://tanstack.com/table/v8/docs/guide/pinning)
-   */
-  setRowPinning: (updater: Updater<RowPinningState>) => void
-}
-
 //
 
 const getDefaultColumnPinningState = (): ColumnPinningState => ({
@@ -255,26 +150,19 @@ const getDefaultColumnPinningState = (): ColumnPinningState => ({
   right: [],
 })
 
-const getDefaultRowPinningState = (): RowPinningState => ({
-  top: [],
-  bottom: [],
-})
-
-export const Pinning: TableFeature = {
-  getInitialState: (state): ColumnPinningTableState & RowPinningState => {
+export const ColumnPinning: TableFeature = {
+  getInitialState: (state): ColumnPinningTableState => {
     return {
       columnPinning: getDefaultColumnPinningState(),
-      rowPinning: getDefaultRowPinningState(),
       ...state,
     }
   },
 
   getDefaultOptions: <TData extends RowData>(
     table: Table<TData>
-  ): ColumnPinningDefaultOptions & RowPinningDefaultOptions => {
+  ): ColumnPinningDefaultOptions => {
     return {
       onColumnPinningChange: makeStateUpdater('columnPinning', table),
-      onRowPinningChange: makeStateUpdater('rowPinning', table),
     }
   },
 
@@ -352,69 +240,6 @@ export const Pinning: TableFeature = {
     row: Row<TData>,
     table: Table<TData>
   ): void => {
-    row.pin = (position, includeLeafRows, includeParentRows) => {
-      const leafRowIds = includeLeafRows
-        ? row.getLeafRows().map(({ id }) => id)
-        : []
-      const parentRowIds = includeParentRows
-        ? row.getParentRows().map(({ id }) => id)
-        : []
-      const rowIds = new Set([...parentRowIds, row.id, ...leafRowIds])
-
-      table.setRowPinning(old => {
-        if (position === 'bottom') {
-          return {
-            top: (old?.top ?? []).filter(d => !rowIds?.has(d)),
-            bottom: [
-              ...(old?.bottom ?? []).filter(d => !rowIds?.has(d)),
-              ...Array.from(rowIds),
-            ],
-          }
-        }
-
-        if (position === 'top') {
-          return {
-            top: [
-              ...(old?.top ?? []).filter(d => !rowIds?.has(d)),
-              ...Array.from(rowIds),
-            ],
-            bottom: (old?.bottom ?? []).filter(d => !rowIds?.has(d)),
-          }
-        }
-
-        return {
-          top: (old?.top ?? []).filter(d => !rowIds?.has(d)),
-          bottom: (old?.bottom ?? []).filter(d => !rowIds?.has(d)),
-        }
-      })
-    }
-    row.getCanPin = () => {
-      const { enableRowPinning, enablePinning } = table.options
-      if (typeof enableRowPinning === 'function') {
-        return enableRowPinning(row)
-      }
-      return enableRowPinning ?? enablePinning ?? true
-    }
-    row.getIsPinned = () => {
-      const rowIds = [row.id]
-
-      const { top, bottom } = table.getState().rowPinning
-
-      const isTop = rowIds.some(d => top?.includes(d))
-      const isBottom = rowIds.some(d => bottom?.includes(d))
-
-      return isTop ? 'top' : isBottom ? 'bottom' : false
-    }
-    row.getPinnedIndex = () => {
-      const position = row.getIsPinned()
-      if (!position) return -1
-
-      const visiblePinnedRowIds = table
-        ._getPinnedRows(position)
-        ?.map(({ id }) => id)
-
-      return visiblePinnedRowIds?.indexOf(row.id) ?? -1
-    }
     row.getCenterVisibleCells = memo(
       () => [
         row._getAllVisibleCells(),
@@ -506,68 +331,6 @@ export const Pinning: TableFeature = {
         return allColumns.filter(d => !leftAndRight.includes(d.id))
       },
       getMemoOptions(table.options, 'debugColumns', 'getCenterLeafColumns')
-    )
-
-    table.setRowPinning = updater => table.options.onRowPinningChange?.(updater)
-
-    table.resetRowPinning = defaultState =>
-      table.setRowPinning(
-        defaultState
-          ? getDefaultRowPinningState()
-          : table.initialState?.rowPinning ?? getDefaultRowPinningState()
-      )
-
-    table.getIsSomeRowsPinned = position => {
-      const pinningState = table.getState().rowPinning
-
-      if (!position) {
-        return Boolean(pinningState.top?.length || pinningState.bottom?.length)
-      }
-      return Boolean(pinningState[position]?.length)
-    }
-
-    table._getPinnedRows = memo(
-      position => [
-        table.getRowModel().rows,
-        table.getState().rowPinning[position!],
-        position,
-      ],
-      (visibleRows, pinnedRowIds, position) => {
-        const rows =
-          table.options.keepPinnedRows ?? true
-            ? //get all rows that are pinned even if they would not be otherwise visible
-              //account for expanded parent rows, but not pagination or filtering
-              (pinnedRowIds ?? []).map(rowId => {
-                const row = table.getRow(rowId, true)
-                return row.getIsAllParentsExpanded() ? row : null
-              })
-            : //else get only visible rows that are pinned
-              (pinnedRowIds ?? []).map(
-                rowId => visibleRows.find(row => row.id === rowId)!
-              )
-
-        return rows
-          .filter(Boolean)
-          .map(d => ({ ...d, position })) as Row<TData>[]
-      },
-      getMemoOptions(table.options, 'debugRows', '_getPinnedRows')
-    )
-
-    table.getTopRows = () => table._getPinnedRows('top')
-
-    table.getBottomRows = () => table._getPinnedRows('bottom')
-
-    table.getCenterRows = memo(
-      () => [
-        table.getRowModel().rows,
-        table.getState().rowPinning.top,
-        table.getState().rowPinning.bottom,
-      ],
-      (allRows, top, bottom) => {
-        const topAndBottom = new Set([...(top ?? []), ...(bottom ?? [])])
-        return allRows.filter(d => !topAndBottom.has(d.id))
-      },
-      getMemoOptions(table.options, 'debugRows', 'getCenterRows')
     )
   },
 }
