+++ packages/table-core/src/features/ColumnSizing.ts
@@ -2,7 +2,7 @@ import { _getVisibleLeafColumns } from '..'
 import { TableFeature } from '../core/table'
 import { RowData, Column, Header, OnChangeFn, Table, Updater } from '../types'
 import { getMemoOptions, makeStateUpdater, memo } from '../utils'
-import { ColumnPinningPosition } from './Pinning'
+import { ColumnPinningPosition } from './ColumnPinning'
 
 //
