+++ packages/table-core/src/index.ts
@@ -10,7 +10,8 @@ export * from './features/Filters'
 export * from './features/Grouping'
 export * from './features/Ordering'
 export * from './features/Pagination'
-export * from './features/Pinning'
+export * from './features/ColumnPinning'
+export * from './features/RowPinning'
 export * from './features/RowSelection'
 export * from './features/Sorting'
 export * from './features/Visibility'
